Remove untagged private images
Every night, we create a new "latest" image. This ends up leaving a
lot of junk behind in the GitHub package UI for our private image
that is hosted in the GitHub container registry. There's hundreds
of older, never to be used, former "most recent nightly" images.
Each is untagged and won't ever be used again (if it was ever used).

This commit adds a steps in the nightly creation workflow to delete
all untagged 'library-documentation-action-v2' images except the
most recent. We leave one as it is possible that something might
reasonably be in the process of pulling down "latest" while we are
doing the deletion and don't want to whack any layers that might
be needed for that.

This change removes a ton of noise of the GitHub packages UI for
'library-documentation-action-v2' and also stops wasting storage
space on GitHub for what we consider to be "junk images".