Do not remove src files for sub-packages
Fixes the condition where sub-packages do not have the src files generated, and links to them fail with a 404.

A current example of this is in the [ponylang/ssl](https://ponylang.github.io/ssl) documentation, where attempting to view any of the source code in [ssl/net](https://ponylang.github.io/ssl/ssl-net--index/) or [ssl/crypto](https://ponylang.github.io/ssl/ssl-crypto--index/) will result in a 404.

The "removing other sources" section of entrypoint.py removed any sources that did not match the library name.

This changes the behaviour to removing sources that do not match library name *or* library_name + '-'. This is consistent with the "removing other docs" section of entrypoint.py.